requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
werkzeug
itsdangerous
# v2
ffmpeg-python
psycopg2-binary